app/requirements.txt
Flask
Flask-Bootstrap
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
Flask-WTF
Flask-Collect
Flask-HTTPAuth
psycopg2-binary
gunicorn
requests
numpy
redis
rq
